[VPD-1659] make volume forms follow correct validation (#120)
* [VPD-1659] AI assisted - make volume forms follow correct validation

* [VPD-1659] removed big decimal

* [VPD-1659] updated error message to make sense

* [VPD-1659] updated validation rules with temp error message

* [VPD-1659] move common test to mappigs

* [VPD-1659] removed unused val

* [VPD-1659] updated to use new error messages

* [VPD-1659] added new volume service to calculate max volume

* [VPD-1659] added new messages

* [VPD-1659] clean up in range duplication

* [VPD-1659] clean up of code

* [VPD-1659] adding more tests

* [VPD-1659] corrected regex and tests with new limits

* [VPD-1659] updated messages

* [VPD-1659] removed big decimal wrapper

* [VPD-1659] normalise max cap for volume

* [VPD-1659] bring duty suspense in line with volume changes

* [VPD-1659] simplified normalise function

* [VPD-1659] updated regex and added test to cover

+++ app/models/returns/SpoiltVolumeByPeriod.scala
@@ -20,7 +20,7 @@ import models.identifiers.PeriodKey
 import play.api.libs.json.{Json, OFormat}
 
 final case class SpoiltVolumeByPeriod(
-  volume: Int,
+  volume: BigDecimal,
   periodKey: PeriodKey
 )
